Treat min10float and min12int as min16float and min16int in AST (#509)
Currently we had separate built in types for min10float and min12int types that will be promoted to min16float and min16int during codegen time. There can be a potential problem with this such as redefinition of functions with min10float or min16float as function parameters, but Sema cannot recognize their redefinitions. This change is to fix so that we promote these lower min precision types to higher min precision types on AST level so that we can easily handle errors.